+++ frontend/src/components/chat/chat-input.tsx
@@ -6,6 +6,7 @@ import { createThread } from "@/actions/create-thread";
 import { useStore } from "@/lib/store";
 import { ToolSelectionDropdown } from "@/components/chat/tool-selection-dropdown";
 import { Send } from "lucide-react";
+import TextareaAutosize from "react-textarea-autosize";
 import ChatInputLoading from "@/components/chat/chat-input-loading";
 import { convert_tools_to_set } from "@/lib/utils";
 import { OpenUserJourneyButton } from "./user-journey-dialog";
@@ -43,12 +44,13 @@ export function ChatInput({ availableTools }: ChatInputProps) {
     startTransition(() => querySuggestions(suggestionInput));
   };
 
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
     if (e.key === "Enter") {
-      e.preventDefault();
-      if (!e.shiftKey) {
-        e.currentTarget.form?.requestSubmit();
+      if (e.shiftKey) {
+        return;
       }
+      e.preventDefault();
+      e.currentTarget.form?.requestSubmit();
     }
   };
   useEffect(() => {
@@ -57,7 +59,7 @@ export function ChatInput({ availableTools }: ChatInputProps) {
   }, []); // Empty dependency array means this runs once on mount
 
   return !isPending ? (
-    <div className="flex flex-col items-center gap-4 pl-2 pr-2">
+    <div className="flex flex-col items-center gap-4 h-[100%] justify-center m-5">
       <h1 className="text-2xl font-bold mt-4 mb-6">
         What can I help you with?
       </h1>
@@ -71,50 +73,46 @@ export function ChatInput({ availableTools }: ChatInputProps) {
           }
           setInput("");
         }}
-        className="w-full flex justify-center"
+        className="flex flex-col w-full max-w-[1200px] justify-center"
       >
-        <div className="min-w-[70%] max-w-[100%]">
-          <div className="flex items-center border-2  border-gray-500 rounded-full overflow-hidden">
-            <input
-              name="content"
-              type="text"
-              autoComplete="off"
-              className="flex-grow p-4 outline-none bg-transparent"
-              placeholder={
-                isPending ? "Creating thread..." : "Message the AI..."
-              }
-              value={input}
-              onChange={(e) => setInput(e.target.value)}
-              onKeyDown={handleKeyDown}
-              disabled={isPending}
-            />
-            <div className="flex gap-2 mr-3">
-              <OpenUserJourneyButton
-                querySuggestions={suggestionActionWrapper}
-                pendingSuggestions={pendingSuggestions}
-              />
-              <ToolSelectionDropdown
-                availableTools={availableTools}
-                checkedTools={checkedTools}
-                setCheckedTools={setCheckedTools}
-              />
-              {isPending ? (
-                <div
-                  className="w-6 h-6 border-2 ml-2 p-1 border-gray-500 border-t-transparent rounded-full animate-spin"
-                  data-testid="loading-spinner"
-                />
-              ) : (
-                <button type="submit" data-testid="send-button" className="p-1">
-                  <Send className="opacity-50" />
-                </button>
-              )}
-            </div>
-          </div>
-          <QuestionSuggestionCards
-            suggestions={suggestionsState as SuggestedQuestions}
-            onSubmit={actionWrapper}
+        <div className="flex items-center border-2  border-gray-500 rounded-[3vw] overflow-hidden min-h-16 pl-9 pr-2">
+          <TextareaAutosize
+            name="content"
+            autoComplete="off"
+            className="flex-grow outline-none border-none bg-transparent resize-none"
+            placeholder={isPending ? "Creating thread..." : "Message the AI..."}
+            value={input}
+            onChange={(e) => setInput(e.target.value)}
+            onKeyDown={(e) => handleKeyDown(e)}
+            disabled={isPending}
+            maxRows={10}
           />
+          <div className="flex gap-3 mr-3">
+            <OpenUserJourneyButton
+              querySuggestions={suggestionActionWrapper}
+              pendingSuggestions={pendingSuggestions}
+            />
+            <ToolSelectionDropdown
+              availableTools={availableTools}
+              checkedTools={checkedTools}
+              setCheckedTools={setCheckedTools}
+            />
+            {isPending ? (
+              <div
+                className="w-6 h-6 border-2 ml-2 p-1 border-gray-500 border-t-transparent rounded-full animate-spin"
+                data-testid="loading-spinner"
+              />
+            ) : (
+              <button type="submit" data-testid="send-button" className="p-1">
+                <Send className="opacity-50" />
+              </button>
+            )}
+          </div>
         </div>
+        <QuestionSuggestionCards
+          suggestions={suggestionsState as SuggestedQuestions}
+          onSubmit={actionWrapper}
+        />
       </form>
     </div>
   ) : (