+++ frontend/src/components/chat/chat-message-ai.tsx
@@ -22,22 +22,22 @@ export const ChatMessageAI = function ChatMessageAI({
     <div className="flex justify-start mt-4">
       {hasTools ? (
         <Button
-          className="hover:scale-105 active:scale-[1.10] ml-7 mr-1 mt-0.5 bg-blue-500 rounded-full p-2.5"
+          className="hover:scale-105 active:scale-[1.10] ml-8 mt-1 bg-blue-500 rounded-full p-2.5"
           onClick={toggleCollapse}
         >
           {toolsCollapsed ? <Wrench /> : <ChevronDown />}
         </Button>
       ) : (
-        <Button className="ml-7 mr-1 mt-0.5 bg-blue-500 rounded-full p-2.5 hover:bg-blue-500">
+        <Button className="ml-8 mt-1 bg-blue-500 rounded-full p-2.5 hover:bg-blue-500">
           <LoaderPinwheel />
         </Button>
       )}
 
       <Card className="max-w-[70%] bg-transparent shadow-none border-none mt-1">
         <CardContent>
-          <span className="prose text-lg pt-8 text-left dark:prose-invert">
+          <div className="prose max-w-none text-lg pt-1 text-left dark:prose-invert">
             <MemoizedMarkdown content={content || ""} id={messageId} />
-          </span>
+          </div>
         </CardContent>
       </Card>
     </div>