+++ frontend/src/components/chat/chat-message-human.tsx
@@ -11,7 +11,7 @@ export const ChatMessageHuman = function ChatMessageHuman({
     <div className="flex justify-end p-8 border-solid break-all">
       <Card className="max-w-[70%]">
         <CardContent>
-          <h1 className="text-lg pt-8">{content}</h1>
+          <h1 className="text-lg pt-8 whitespace-pre-wrap">{content}</h1>
         </CardContent>
       </Card>
     </div>