+++ frontend/src/lib/fetcher.ts
@@ -1,5 +1,5 @@
 import { env } from "@/lib/env";
-
+import { CustomError } from "@/lib/types";
 const isServer = typeof window === "undefined";
 
 type Method = "GET" | "PATCH" | "POST" | "PUT" | "DELETE";
@@ -64,8 +64,9 @@ export async function fetcher({
   });
 
   if (!response.ok) {
-    throw new Error(
-      `Fetching Error : ${response.status}, ${response.statusText}`,
+    throw new CustomError(
+      `Fetching Error : ${response.statusText}`,
+      response.status,
     );
   }
   return await response.json();