+++ CHANGELOG.md
@@ -88,6 +88,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Fix docker compose redirect bug.
 - [frontend] - Refactor
 - Websearch tools.
+- Small bugfixes.
+- [frontend] - Fix suggestion UI.
+- [frontend] - cosmetic fixes.
 
 ## [0.3.3] - 30.10.2024